+++ CHANGES.rst
@@ -26,6 +26,9 @@ Changelog
 - Some tuning of migration from PFG
   [ThibautBorn]
 
+- Test max_size validator
+  [gotcha]
+
 
 4.3.0 (2024-12-13)
 ------------------